[librptexture] Use signed iterators for OpenMP `for` loops.
MSVC 2015 implements OpenMP 2.0, which only supports integral types
for `for` loop iterators. OpenMP 3.0 supports this, and gcc-11.2.0
implements OpenMP 4.5 (and partial OpenMP 5.0).

Interestingly, switching to `int` slighly reduced the code size in
the gcc-11.2.0 build:

Code size difference: (64-bit Gentoo Linux, gcc-11.2.0, release build, no LTO)

   text    data     bss     dec     hex filename
   4668       8       0    4676    1244 ImageDecoder_BC7.cpp.o [unsigned iterator]
   4652       8       0    4660    1234 ImageDecoder_BC7.cpp.o [signed iterator]
    -16       0       0     -16     -10 Difference

   text    data     bss     dec     hex filename
   1657       8       0    1665     681 ImageDecoder_ASTC.cpp.o [unsigned iterator]
   1641       8       0    1649     671 ImageDecoder_ASTC.cpp.o [signed iterator]
    -16       0       0     -16     -10 Difference

+++ src/librptexture/decoder/ImageDecoder_ASTC.cpp
@@ -71,8 +71,8 @@ rp_image *fromASTC(int width, int height,
 	// so we'll need to use a tiled decode loop.
 
 	// Calculate the total number of tiles.
-	const unsigned int tilesX = static_cast<unsigned int>(physWidth / block_x);
-	const unsigned int tilesY = static_cast<unsigned int>(physHeight / block_y);
+	const int tilesX = physWidth / block_x;
+	const int tilesY = physHeight / block_y;
 	const unsigned int bytesPerTileRow = tilesX * 16;	// for OpenMP
 
 	// NOTE: Largest ASTC format is 12x12.
@@ -84,9 +84,9 @@ rp_image *fromASTC(int width, int height,
 #endif /* _OPENMP */
 
 #pragma omp parallel for
-	for (unsigned int y = 0; y < tilesY; y++) {
+	for (int y = 0; y < tilesY; y++) {
 		const uint8_t *pSrc = &img_buf[y * bytesPerTileRow];
-		for (unsigned int x = 0; x < tilesX; x++, pSrc += 16) {
+		for (int x = 0; x < tilesX; x++, pSrc += 16) {
 			// Temporary tile buffer
 			array<uint32_t, 12*12> tileBuf;
